backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.6
redis==5.0.1

# WebSocket
python-socketio==5.11.0
python-engineio==4.9.0

# PDF Processing
pdfplumber==0.10.4
pypdf2==3.0.1

# Word Document Generation
python-docx==1.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client (for testing)
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Monitoring & Logging
sentry-sdk==1.39.2

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
